requirements.txt
# ─── Core Django ──────────────────────────────────────────────────────────────
Django==4.2.11
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1

# ─── PostgreSQL driver ────────────────────────────────────────────────────────
psycopg2-binary==2.9.9

# ─── Utilities ───────────────────────────────────────────────────────────────
django-cors-headers==4.3.1
Pillow==10.3.0
python-dotenv==1.0.1

# ─── Removed (MongoDB / djongo) ───────────────────────────────────────────────
# djongo==1.3.6         ← removed
# pymongo==3.12.3       ← removed
# sqlparse==0.2.4       ← was required by djongo; no longer needed
